+++ src/lib/middleware/index.ts
@@ -1,0 +1,9 @@
+import { RequestHandler } from "src/types";
+
+interface MiddlewareObj {
+  handler: RequestHandler;
+}
+
+export function defineMiddlewares(middlewares: MiddlewareObj[]) {
+  return middlewares.map((m) => m.handler);
+}